+++ ToDoAppANDROID/app/src/main/java/com/example/todoappandroid/TaskEntity.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "tasks")
 data class TaskEntity(
     @PrimaryKey
-    val id: Int = System.currentTimeMillis().toInt(),
+    val id: Long = System.currentTimeMillis(),
     val title: String,
     val description: String? = null,
     val date: String? = null,
